pyproject.toml
[build-system]
requires = [
    'setuptools==68.1.2',
]
build-backend = 'setuptools.build_meta'

[project]
name = 'zametka'
version = '0.0.1'
description = 'zametka API'
readme = 'README.md'
requires-python = '>=3.11.1'
dependencies = [
    'alembic==1.11.1',
    'asyncpg==0.29.0',
    'fastapi==0.111.0',
    'SQLAlchemy==2.0.22',
    'uvicorn==0.29.0',
    'email-validator==2.1.1',
    'argon2-cffi',
    'aiosmtplib==2.0.2',
    'Jinja2~=3.1.2',
    'pydantic~=2.7.1',
    'starlette~=0.37.2',
    'aiohttp==3.9.1',
    'Brotli==1.1.0',
    'adaptix==3.0.0b5',
    'aio-pika==9.3.1',
    'dishka==1.1.1',
    'PyJWT==2.8.0',
]

[project.optional-dependencies]
dev = [
    'pytest==8.2.0',
    'pytest-asyncio==0.23.6',
    'mypy==1.3.0',
    'ruff==0.4.2',
    'pre-commit==3.7.0',
]
test = [
    'pytest==8.2.0',
    'pytest-asyncio==0.23.6',
]
lint = [
    'mypy==1.3.0',
    'ruff==0.4.2'
]


[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    'access: tests related to access context.',
    'notes: tests related to notes context.',
    'domain: domain tests',
    'application: application tests',
]
filterwarnings = "ignore::DeprecationWarning"

[project.scripts]
zametka = "zametka.main.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.mypy]
strict = true
warn_unreachable = true
show_column_numbers = true
show_error_context = true
check_untyped_defs = true
ignore_missing_imports = false

[tool.ruff]
line-length = 88
exclude = [
    "notes"
]

[tool.ruff.lint]
select = ['ALL']

ignore = [
  # Rules that should be turned on in the near future
  'N818',
  'B904',
  'FIX002',
  'RUF012',

  # Rules emitting false alerts
  'N804',
  'B008',
  'BLE001',
  'RUF009',
  'RUF001',

  # Rules that are not applicable in the project for now
  'D',
  'TID252',
  'D104',
  'ANN',
  'SLF001',
  'ARG',
  'D100',
  'PLR0913',
  'TCH002',
  'EXE002',

  # Strange and obscure rules that will never be turned on
  'ANN101',
  'FA100',
  'TRY003',
  'TRY201',
  'EM',
  'PERF203',
  'TCH001',
  'TD002',
  'PTH201',
  'RSE102',
  'RET504',
  'FBT001',
  'TD003',
  'B024',
  'ISC001',
  'B027',
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ['F401']

"test_*" = ['S101', 'PLR2004', 'PT023', 'PT001', 'PT006']
"conftest.py" = ['PT023', 'PT001', 'PT006']
"cli.py" = ["T201"]

[[project.authors]]
name = 'lubaskinc0de'
email = 'lubaskincorporation@gmail.com'
